DFE-390 Add subscriptions page
